finale

+++ TP/new_server/challengesHandler.h
@@ -18,6 +18,6 @@ const char *getChallengeResearchQuestion(t_challengeADT challenge);
 
 void getChallengeFunction(t_challengeADT challenge);
 
-uint8_t isValidAnswer(t_challengeADT challenge, const char *answer);
+int isValidAnswer(t_challengeADT challenge, const char *answer);
 
 #endif
